requirements.txt
discord.py[voice]
beautifulsoup4~=4.9.3
bs4~=0.0.1
colorama
colorlog
colormap~=1.0.3
ffmpeg
lavalink~=3.1.2
more-itertools
Pillow~=7.2.0
prawcore
psutil~=5.7.2
Pybooru~=4.2.0
termcolor~=1.1.0
urllib3
Wand
wikipedia~=1.4.0
youtube-dl
tqdm
aiosqlite~=0.15.0
aiofiles~=0.5.0
aiomysql~=0.0.20
pymysql~=0.9.2
easydev
profanity-check
praw~=7.1.0

requests~=2.24.0
aiohttp~=3.6.2
numpy~=1.19.2
websockets~=8.1
pysftp~=0.2.9